% Generated by roxygen2 (4.0.2): do not edit by hand
\name{carbonContent}
\alias{carbonContent}
\title{Fitted Carbon Content}
\usage{
carbonContent(times, Year, PMC)
}
\arguments{
\item{times}{a fitted object returned by \code{metropolisDate}}

\item{Year}{calibration data - (fractional) year of atmospheric
carbon measurement}

\item{PMC}{calibration data - atmospheric carbon concentration}
}
\value{
a coda object describing the growth rates
}
\description{
Posterior prediction of segment mean carbon content
}
\details{
Calculates the posterior segment mean carbon content given
atmospheric calibration data.
}

