% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{DateMoss-package}
\alias{DateMoss-package}
\title{Chronologies and growth rates of moss}
\description{
Provides facilities for estimating chronologies and growth rates
of moss (or similar) samples based on radiocarbon dating.
}
\author{
B. Raymond, S. Wotherspoon
}

