% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generateInitial}
\alias{generateInitial}
\title{Initial Chronologies}
\usage{
generateInitial(lengths, tmin, tmax, chains = 1, max.growth = NULL)
}
\arguments{
\item{lengths}{a vector of segment lengths (in timewise order)}

\item{tmin}{the minimum possible date}

\item{tmax}{the final date}

\item{chains}{the number of chains to generate initializations for}

\item{max.growth}{the maximum possible growth rate}
}
\value{
a list of vectors of dates
}
\description{
Initial Chronology for initializing MCMC sampler
}
\details{
The function \code{generateInitial} generates random chronologies
that can be used to initialize the MCMC sampler
\code{\link{metropolisDate}}.

The user must provide the vector \code{lengths} containing the
continuous sequence of segment lengths in timewise order, the
earliest possible date \code{tmin} and the final date \code{tmax}
(usually the acquisition date).  If in addition the user specifies
a maximal growth rate then dates will be generated such that the
growth rate never exceeds this upper limit, and if this is not
possible the function will raise an error.
}

