% Generated by roxygen2 (4.0.2): do not edit by hand
\name{growth.prior.uniform}
\alias{growth.prior.gamma}
\alias{growth.prior.uniform}
\title{Growth Priors}
\usage{
growth.prior.uniform(max.growth)

growth.prior.gamma(mean, sd, shape = (mean/sd)^2, rate = mean/sd^2)
}
\arguments{
\item{max.growth}{the maximum possible growth rate}

\item{mean}{the mean of the Gamma prior for growth}

\item{sd}{the standard deviation of the Gamma prior for growth}

\item{shape}{the shape parameter of the Gamma prior for growth}

\item{rate}{the rate parameter of the Gamma prior for growth}
}
\value{
function to evaluate the contribution to the log prior
from each segment.
}
\description{
Priors for growth rates.
}
\details{
These functions construct functions to evaluate the contribution
to the log prior from each segment, for use with
\code{metropolisDate}.

The \code{growth.prior.uniform} creates a uniform prior that
assumes growth rates are uniform up to a user specified maximum
rate.

The \code{growth.prior.gamma} creates a Gamma prior for grwoth
rates.  The user can specify the either the parameters of the
distribution as either the prior mean and standard deviation or
the prior shape and rate of the Gamma distribution.
}

