% Generated by roxygen2 (4.0.2): do not edit by hand
\name{growthRate}
\alias{growthRate}
\title{Estimate Growth Rates}
\usage{
growthRate(times, lengths)
}
\arguments{
\item{times}{a fitted object returned by \code{metropolisDate}}

\item{lengths}{a vector of segment lengths}
}
\value{
a coda object describing the growth rates
}
\description{
Estimate growth rates from the estimated segment times
}
\details{
This function generates samples of segment growth rates from the
sample of segment cut dates generated by \code{metropolisDate}, and
the vector \code{lengths} of segment lengths in timewise order.
}

