% Generated by roxygen2 (4.0.2): do not edit by hand
\name{metropolisDate}
\alias{metropolisDate}
\alias{metropolisDate0}
\title{Estimate segment dates}
\usage{
metropolisDate(ts.init, tmin, len, pmc, Year, PMC, iters = 2000L,
  thin = 25L, chains = if (is.list(ts.init)) length(ts.init) else 1L,
  log.prior = NULL, alpha = 0.01, beta = 0.01, verbose = interactive())

metropolisDate0(ts.init, tmin, len, pmc, Year, PMC, sigma = 4,
  iters = 2000L, thin = 25L, chains = if (is.list(ts.init))
  length(ts.init) else 1L, log.prior = NULL, verbose = interactive())
}
\arguments{
\item{ts.init}{a vector or a list of vectors of initial date
estimates, as generated by \code{generateInitial}}

\item{tmin}{the minimum possible date}

\item{len}{lengths of the segments (in timewise order, oldest
first)}

\item{pmc}{mean percentage modern carbon concentration of the
segments (in timewise order, oldest first)}

\item{Year}{calibration data - (fractional) year of atmospheric
carbon measurement}

\item{PMC}{calibration data - percent modern carbon}

\item{iters}{number of samples to draw.}

\item{thin}{rate at which to thin samples.}

\item{chains}{number of chains to sample.}

\item{log.prior}{function to calculate the contribution from each
segment to the log prior for growth rates, given segment times and
lengths.}

\item{alpha}{shape parameter for the Gamma prior for tau}

\item{beta}{rate parameter for the Gamma prior for tau}

\item{verbose}{report progress at prompt?}

\item{sigma}{standard deviation of PMC errors}
}
\value{
\code{metropolisDate0} returns a list containing a coda
object describing the segment times.  \code{metropolisDate}
returns a list containing a coda object describing the segment
times and a coda object describing
}
\description{
Estimate segment times and growth rates by MCMC
}
\details{
The two functions \code{metropolisDate0} and \code{metropolisDate}
estimate the segment cut times by Metropolis sampling.  The
\code{metropolisDate0} variant assumes the standard deviation of
the PMC errors (\code{sigma}) is fixed, while
\code{metropolisDate} Gibbs samples for \code{sigma}.

The user must supply the lengths \code{len} and mean carbon
concentrations \code{pmc} of the segments.  These must be supplied
in timewise order - oldest segment first and the segments must be
contiguous, but the concentration data may contain missing values.

The user must also supply atmospheric radiocarbon calibration data
as two vectors, \code{Year} the time of measurement in
(fractional) years and \code{PMC} the recorded percentage of
modern carbon.

The user may also supply a function that computes the
contributions of each segment to the log prior.  This function
takes two arguments, \code{ts} a vector of the (n+1) segment
endpoint times and \code{len} the lengths of the n segments.  This
function must compute the log of the (non-normalized) contributions
to prior from each segment.
}

