% Generated by roxygen2 (4.0.2): do not edit by hand
\name{quantile.mcmc}
\alias{quantile.mcmc}
\alias{quantile.mcmc.list}
\title{Quantiles}
\usage{
\method{quantile}{mcmc}(x, ...)

\method{quantile}{mcmc.list}(x, ...)
}
\arguments{
\item{x}{a coda object}

\item{...}{further arguments passed to quantile}
}
\value{
An array of quantiles
}
\description{
Quantiles from a coda object
}
\details{
Converts an \code{mcmc} or \code{mcmc.list} object to a matrix and
applies quantile.
}

