% Generated by roxygen2 (4.0.2): do not edit by hand
\name{smoothCalibration}
\alias{smoothCalibration}
\title{Calibration Curve Smoothing}
\usage{
smoothCalibration(Year, PMC, smoothing.level, method = "rgr")
}
\arguments{
\item{Year}{calibration data - vector of (fractional) year of atmospheric carbon measurement}

\item{PMC}{calibration data - vector of percent modern carbon values}

\item{smoothing.level}{(scalar) the level of smoothing to apply. The interpretation of this value depends on the method used}

\item{method}{string specifying the smoothing method to use. Currently only "rgr"}
}
\value{
a data frame of smoothed carbon values with column names "Year" and "PMC"
}
\description{
Smooth calibration curve
}
\details{
The function \code{smoothCalibration} smooths
an atmospheric carbon curve. This function is experimental.

A smoothed calibration curve can be used to provide an approximation for the
effect of having multiple strands in a sample of moss, each with
slightly different growth rates or physical alignment within the sample.
Currently the only method provided is "rgr", which
effectively assumes that the sample is comprised of multiple strands, each
with a constant (random) growth rate offset.
}

