% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{Atmospheric14C}
\alias{Atmospheric14C}
\title{Atmospheric radiocarbon calibration data}
\format{A data frame with 6 columns and 163 rows.  The columns
represent
\tabular{rl}{
\code{Year} \tab Year of measurement. \cr
\code{F14C} \tab FIX ME. \cr
\code{F14C.sigma} \tab FIX ME. \cr
\code{D14C} \tab FIX ME. \cr
\code{F14C.sigma} \tab FIX ME. \cr
\code{PMC} \tab Percent modern carbon.
}}
\source{
Hua, Q., Barbetti, M. & Rakowski, A. Z. (2013).
Atmospheric radiocarbon for the period 1950-2010.
Radiocarbon, 55(4), 2059-2072.
}
\description{
Atmospheric radiocarbon calibration data
}
\details{
Atmospheric radiocarbon data from 1450-2013.  These data are
largely based on Hua et al (2013), but have been extrapolated at
both ends.
}
\keyword{data}

